+++ linkerd/addr/fuzz/Cargo.toml
@@ -11,9 +11,8 @@ cargo-fuzz = true
 [target.'cfg(fuzzing)'.dependencies]
 libfuzzer-sys = "0.4"
 linkerd-addr = { path = ".." }
+linkerd-tracing = { path = "../tracing" }
 tracing = "0.1.23"
-tracing-subscriber = "0.2.17"
-
 # Prevent this from interfering with workspaces
 [workspace]
 members = ["."]
