+++ linkerd/app/inbound/fuzz/fuzz_targets/fuzz_target_1.rs
@@ -5,7 +5,9 @@ use {libfuzzer_sys::fuzz_target, linkerd_app_inbound::http::fuzz_logic::*};
 
 #[cfg(fuzzing)]
 fuzz_target!(|requests: Vec<HttpRequestSpec>| {
-    let _ = tracing_subscriber::fmt::try_init();
+    // Don't enable tracing in `cluster-fuzz`, since we would emit verbose
+    // traces for *every* generated fuzz input...
+    let _trace = linkerd_tracing::test::with_default_filter("off");
     tracing::info!(?requests, "running with input");
     if requests.len() == 0 {
         return;
