+++ linkerd/app/outbound/Cargo.toml
@@ -35,4 +35,4 @@ linkerd-app-test = { path = "../test" }
 linkerd-io = { path = "../../io", features = ["tokio-test"] }
 tokio = { version = "1", features = ["full", "macros"]}
 tokio-test = "0.4"
-tracing-subscriber = "0.2.16"
+linkerd-tracing = { path = "../../tracing", features = ["ansi"] }
