+++ linkerd/app/test/src/lib.rs
@@ -13,8 +13,6 @@ pub use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 pub use tokio::sync::oneshot;
 pub use tower::Service;
 pub use tracing::*;
-pub use tracing_subscriber::prelude::*;
-
 pub mod io {
     pub use tokio::io::*;
     pub use tokio_test::io::*;
@@ -49,34 +47,6 @@ pub struct ContextError {
     context: Cow<'static, str>,
     source: Error,
     at: &'static Location<'static>,
-}
-
-/// By default, disable logging in modules that are expected to error in tests.
-const DEFAULT_LOG: &str = "warn,\
-                           linkerd=debug,\
-                           linkerd_proxy_http=error,\
-                           linkerd_proxy_transport=error";
-
-pub fn trace_subscriber() -> (Dispatch, app_core::trace::Handle) {
-    use std::env;
-    let log_level = env::var("LINKERD2_PROXY_LOG")
-        .or_else(|_| env::var("RUST_LOG"))
-        .unwrap_or_else(|_| DEFAULT_LOG.to_string());
-    let log_format = env::var("LINKERD2_PROXY_LOG_FORMAT").unwrap_or_else(|_| "PLAIN".to_string());
-    env::set_var("LINKERD2_PROXY_LOG_FORMAT", &log_format);
-    // This may fail, since the global log compat layer may have been
-    // initialized by another test.
-    let _ = app_core::trace::init_log_compat();
-    app_core::trace::Settings::default()
-        .filter(log_level)
-        .format(log_format)
-        .test(true)
-        .build()
-}
-
-pub fn trace_init() -> tracing::dispatcher::DefaultGuard {
-    let (d, _) = trace_subscriber();
-    tracing::dispatcher::set_default(&d)
 }
 
 // === impl ContextError ===
