+++ linkerd/dns/fuzz/Cargo.toml
@@ -14,7 +14,7 @@ libfuzzer-sys = "0.4"
 linkerd-dns = { path = ".." }
 tokio = { version = "1", features = ["rt", "time", "io-util"] }
 tracing = "0.1.23"
-tracing-subscriber = "0.2.17"
+linkerd-tracing = { path = "../tracing" }
 
 # Prevent this from interfering with workspaces
 [workspace]
