+++ linkerd/tls/Cargo.toml
@@ -28,6 +28,6 @@ untrusted = "0.7"
 [dev-dependencies]
 linkerd-identity = { path = "../identity", features = ["test-util"] }
 linkerd-proxy-transport = { path = "../proxy/transport" }
+linkerd-tracing = { path = "../tracing", features = ["ansi"]}
 tokio = { version = "1", features = ["rt-multi-thread"] }
 tower = { version = "0.4.7", default-features = false, features = ["util"] }
-tracing-subscriber = "0.2.16"
