+++ linkerd/tracing/src/lib.rs
@@ -2,6 +2,7 @@
 
 mod level;
 mod tasks;
+pub mod test;
 mod uptime;
 
 use self::uptime::Uptime;
@@ -123,8 +124,7 @@ impl Settings {
                     .event_format(fmt)
                     // Since we're using the JSON event formatter, we must also
                     // use the JSON field formatter.
-                    .fmt_fields(format::JsonFields::default())
-                    .with_ansi(false);
+                    .fmt_fields(format::JsonFields::default());
                 let registry = registry.with(tasks_layer);
                 let dispatch = if self.test {
                     registry.with(fmt.with_test_writer()).into()
@@ -136,9 +136,7 @@ impl Settings {
             _ => {
                 let (tasks, tasks_layer) = TasksLayer::<format::DefaultFields>::new();
                 let registry = registry.with(tasks_layer);
-                let fmt = tracing_subscriber::fmt::layer()
-                    .event_format(fmt)
-                    .with_ansi(self.test);
+                let fmt = tracing_subscriber::fmt::layer().event_format(fmt);
                 let dispatch = if self.test {
                     registry.with(fmt.with_test_writer()).into()
                 } else {
