+++ linkerd/transport-header/fuzz/Cargo.toml
@@ -14,8 +14,7 @@ libfuzzer-sys = { version = "0.4.0", features = ["arbitrary-derive"] }
 linkerd-transport-header = { path = ".." }
 tokio = { version = "1", features = ["full"] }
 tracing = "0.1.23"
-tracing-subscriber = "0.2.17"
-
+linerd-tracing = { path = "../tracing" }
 # Prevent this from interfering with workspaces
 [workspace]
 members = ["."]
